test: update tests for ProductCard component
