+++ src/components/ProductCard/__test__/ProductCard.test.tsx
@@ -1,7 +1,21 @@
 import { cleanup } from '@testing-library/react';
 import { afterEach, describe, expect, it } from 'vitest';
 
-import { mockBook, setup } from './helper';
+import ProductCard from '../ProductCard';
+import { renderWithProviders } from '@/utils/test-utils';
+
+export const mockBook = {
+  authors: ['Автор тест1', 'Автор тест2'],
+  id: 1,
+  title: 'Тестова назва книги',
+  imageUrl: 'тестова картинка',
+  price: 100,
+  discount: 10,
+  discountPrice: 90,
+  year: 2020,
+  totalQuantity: 100,
+  expected: true,
+};
 
 describe('ProductCard Component', () => {
   afterEach(() => {
@@ -9,7 +23,9 @@ describe('ProductCard Component', () => {
   });
 
   it('renders the ProductCard with title, price, and discount price', () => {
-    const { getByText } = setup();
+    const { getByText } = renderWithProviders(
+      <ProductCard data={mockBook} variant="catalog" />
+    );
 
     expect(getByText(mockBook.title)).toBeInTheDocument();
     expect(getByText(mockBook.price)).toBeInTheDocument();
@@ -18,7 +34,9 @@ describe('ProductCard Component', () => {
 
   it('renders the ProductCard without discount price when discount is zero', () => {
     const book = { ...mockBook, discount: 0, discountPrice: 0 };
-    const { getByText, queryByText } = setup({ data: book });
+    const { getByText, queryByText } = renderWithProviders(
+      <ProductCard data={book} variant="catalog" />
+    );
 
     expect(getByText(mockBook.title)).toBeInTheDocument();
     expect(getByText(mockBook.price)).toBeInTheDocument();
